\name{awesomeakshathyaning-package}
\alias{awesomeakshathyaning-package}
\alias{awesomeakshathyaning}
\docType{package}
\title{
\packageTitle{awesomeakshathyaning}
}
\description{
\packageDescription{awesomeakshathyaning}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{awesomeakshathyaning}
\packageIndices{awesomeakshathyaning}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{awesomeakshathyaning}

Maintainer: \packageMaintainer{awesomeakshathyaning}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
