% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Finding The Highest Common Factor Of Two Numbers}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{one of the number to compute HCF}

\item{y}{another number to compute HCF}
}
\value{
returns HCF of two numbers
}
\description{
Finding The Highest Common Factor Of Two Numbers
}
